move helper functions into helpers from notifications, fix researcher comments
