+++ helpers/forms.py
@@ -34,3 +34,11 @@ UpdateTKLabelTranslationFormSet = inlineformset_factory(
         'translation': forms.Textarea(attrs={'class': 'w-100 margin-top-1 margin-bottom-2', 'style': 'height: 150px; padding: 10px;'}),
     }
 )
+
+class NoticeCommentForm(forms.ModelForm):
+    class Meta:
+        model = NoticeComment
+        fields = ['message']
+        widgets = {
+            'message': forms.TextInput(attrs={'class': 'w-100', 'placeholder': 'Add Comment'})
+        }
